backend/app/requirements.txt
flask==3.0.2
requests==2.32.5
python-dotenv==1.0.0
flask-cors==6.0.2
